Songs endpoint

+++ src/main/java/com/protify/Protify/repository/SongRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SongRepository extends JpaRepository<Songs, Integer> {}
+public interface SongRepository extends JpaRepository<Songs, Long> {}
